refactor: major refactor of the layout, and code organization
